[FE][fix] labelList 컴포넌트 분리
labelList page, component 분리
