+++ frontend/components/label/LabelList.jsx
@@ -1,0 +1,97 @@
+import React, { useEffect } from 'react';
+import Mytable from '@components/common/table';
+import service from '@services';
+import Label from '@components/common/label';
+import fontColorContrast from 'font-color-contrast';
+import styled from 'styled-components';
+import PropTypes from 'prop-types';
+
+const LabelList = ({ data, getData }) => {
+  const clickEditBtn = async (label) => {};
+  const clickDeleteBtn = async (id) => {
+    await service.deleteLabel(id);
+    await getData();
+  };
+
+  return (
+    <>
+      {/* {data} */}
+      <Mytable
+        width='100%'
+        header={() => {
+          return (
+            <LabelListHeader>
+              <td colSpan='5'>
+                <HeaderText>{data.length}labels</HeaderText>{' '}
+              </td>
+            </LabelListHeader>
+          );
+        }}
+        body={() => {
+          return data.map((label) => {
+            return (
+              <TR key={label.id}>
+                <TD width='200px;' align='left' padding='10px'>
+                  <Label text={label.title} bg={label.color} color={fontColorContrast(label.color)} />
+                </TD>
+                <TD align='left'>{label.description}</TD>
+
+                <TD>
+                  <BTN onClick={() => clickEditBtn(label)}>edit</BTN>
+                </TD>
+                <TD>
+                  <BTN onClick={() => clickDeleteBtn(label.id)}>delete</BTN>
+                </TD>
+              </TR>
+            );
+          });
+        }}
+      />
+    </>
+  );
+};
+LabelList.propTypes = {
+  data: PropTypes.array.isRequired,
+  getData: PropTypes.func.isRequired,
+};
+
+const LabelListHeader = styled.tr`
+  background-color: #eee;
+  /* display: flex; */
+  /* justify-content: center; */
+  height: 60px;
+  padding-left: 18px;
+  /* align-items: center; */
+`;
+
+const TR = styled.tr`
+  border-bottom: 1px solid #eee;
+`;
+
+const TD = styled.td`
+  width: ${(props) => props.width || ''};
+  text-align: ${(props) => props.align || 'center'};
+  padding: ${(props) => props.padding || '5px'};
+`;
+
+const HeaderText = styled.p`
+  color: gray;
+  padding-left: 20px;
+  font-size: 14px;
+`;
+
+const BTN = styled.button`
+  width: ${(props) => props.width || ''};
+  text-align: ${(props) => props.align || 'center'};
+  padding: ${(props) => props.padding || '5px'};
+  border: none;
+  cursor: pointer;
+  background: none;
+  color: grey;
+  padding: 1px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+`;
+export default LabelList;
